refactor: move script to root of .war file
WebContent/ is root of the .war file

+++ WebContent/log_processing.py
@@ -19,6 +19,6 @@ def measure_time():
 
 def parse_line(line):
     ts, tj = line.split(",")
-    return int(ts[3:]), int(tj[3:])
+    return float(ts[3:]), float(tj[3:])
 
 measure_time()
